Anonymous channels instead of object identity.
I realized creating a keypair is very heavyweight, and is basically an
object identity primitive. Object identity is useful, but channels are
probably better done with discovery so that they are lightweight and
can be created in bulk.

Keypairs are probably better for entire machines: a node will only
accept messages on a channel signed by a public key it trusts, rather
than just anyone who knew the symmetric key.